reorganize BizShow/MainContent
