+++ frontend/src/components/BusinessShow/MainContent/share-modal.css
@@ -20,12 +20,11 @@
 
 .share-modal-content {
   position: absolute;
-  /* background-color: white; */
   background: #f1f1f1;
-  top: 10px;
+  /* top: 10px; */
+  /* bottom: 10px; */
   display: flex;
   border-radius: 10px;
-  width: 50vw;
-  height: 90vh;
-  /* z-index: 2; */
+  width: 700px;
+  height: 98.5vh;
 }
